start convert softseg new file
